+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandFreezeExtractionConfiguration.cs
@@ -5,19 +5,17 @@
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
 using System.Drawing;
-using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.DataExport.Data;
 using Rdmp.Core.Icons.IconProvision;
-using Rdmp.UI.ItemActivation;
 using ReusableLibraryCode.Icons.IconProvision;
 
-namespace Rdmp.UI.CommandExecution.AtomicCommands
+namespace Rdmp.Core.CommandExecution.AtomicCommands
 {
-    internal class ExecuteCommandFreezeExtractionConfiguration : BasicUICommandExecution, IAtomicCommand
+    internal class ExecuteCommandFreezeExtractionConfiguration : BasicCommandExecution, IAtomicCommand
     {
         private ExtractionConfiguration _extractionConfiguration;
 
-        public ExecuteCommandFreezeExtractionConfiguration(IActivateItems activator, ExtractionConfiguration extractionConfiguration) : base(activator)
+        public ExecuteCommandFreezeExtractionConfiguration(IBasicActivateItems activator, ExtractionConfiguration extractionConfiguration) : base(activator)
         {
             _extractionConfiguration = extractionConfiguration;